requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
PyMySQL==1.1.0
reportlab==4.4.5
Flask-Mail==0.9.1
PyJWT[crypto]==2.10.1
